Refactoring Codebase
