+++ build.gradle
@@ -6,20 +6,14 @@ plugins {
     id 'org.openjfx.javafxplugin' version '0.0.13'
 }
 
-javafx {
-    version = "19"
-    modules = [ 'javafx.controls' ]
-}
-
-checkstyle {
-    toolVersion = '10.2'
-}
-
 repositories {
     mavenCentral()
 }
 
 dependencies {
+    testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: '5.5.0'
+    testRuntimeOnly group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.5.0'
+
     String javaFxVersion = '11'
 
     implementation group: 'org.openjfx', name: 'javafx-base', version: javaFxVersion, classifier: 'win'
@@ -34,21 +28,6 @@ dependencies {
     implementation group: 'org.openjfx', name: 'javafx-graphics', version: javaFxVersion, classifier: 'win'
     implementation group: 'org.openjfx', name: 'javafx-graphics', version: javaFxVersion, classifier: 'mac'
     implementation group: 'org.openjfx', name: 'javafx-graphics', version: javaFxVersion, classifier: 'linux'
-}
-
-jar {
-    manifest {
-        attributes(
-
-                'Main-Class': 'Launcher',
-        )
-    }
-}
-
-dependencies {
-    testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: '5.5.0'
-    testRuntimeOnly group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.5.0'
-    compile group: 'com.joestelmach', name: 'natty', version: '0.6'
 }
 
 test {
@@ -74,7 +53,10 @@ shadowJar {
     archiveClassifier = null
 }
 
+checkstyle {
+    toolVersion = '8.29'
+}
+
 run{
     standardInput = System.in
-    enableAssertions = true
 }
