+++ src/main/java/tasklist/TaskList.java
@@ -1,10 +1,17 @@
-package duke;
+package tasklist;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+
+import exceptions.DukeException;
+import tasks.Deadline;
+import tasks.Event;
+import tasks.PeriodTask;
+import tasks.Task;
+import tasks.Todo;
 
 /**
  * Representation of the list containing tasks
