+++ src/main/java/tasks/Task.java
@@ -1,4 +1,4 @@
-package duke;
+package tasks;
 
 /**
  * Class of task
