+++ src/test/java/duke/TodoTest.java
@@ -1,6 +1,10 @@
 package duke;
 
 import org.junit.jupiter.api.Test;
+
+import tasks.Task;
+import tasks.Todo;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class TodoTest {
